I'm adding touch support

+++ include/Events/Mouse.hpp
@@ -4,12 +4,12 @@
 
 struct MouseButton
 {
-    MouseButton(const SDL_MouseButtonEvent* ev, uint8_t ratio):
+    MouseButton(const SDL_MouseButtonEvent* ev, float ratio):
         pressed(ev->state),
         id(ev->which),
         button(ev->button),
         clicks(ev->clicks),
-        position((float)ev->x / ratio, (float)ev->y / ratio)
+        position(ev->x / ratio, ev->y / ratio)
     {}
 
     const bool pressed;
@@ -22,11 +22,11 @@ struct MouseButton
 
 struct MouseMove
 {
-    MouseMove(const SDL_MouseMotionEvent* ev, uint8_t ratio) :
+    MouseMove(const SDL_MouseMotionEvent* ev, float ratio) :
         id(ev->which),
         state(ev->state),
-        position((float)ev->x / ratio, (float)ev->y / ratio),
-        delta((float)ev->xrel / ratio, (float)ev->yrel / ratio)
+        position(ev->x / ratio, ev->y / ratio),
+        delta(ev->xrel / ratio, ev->yrel / ratio)
     { }
 
     const Uint32 id;
@@ -38,11 +38,11 @@ struct MouseMove
 
 struct MouseWheel
 {
-    MouseWheel(const SDL_MouseWheelEvent* ev, uint8_t ratio):
+    MouseWheel(const SDL_MouseWheelEvent* ev, float ratio):
         id(ev->which),
         amount(ev->x, ev->y),
         p_amount(ev->preciseX, ev->preciseY),
-        position((float)ev->mouseX / ratio, (float)ev->mouseY / ratio)
+        position(ev->mouseX / ratio, ev->mouseY / ratio)
     {}
 
     const Uint32 id;
